requirements.txt
Flask
torch
transformers
Pillow
PyMuPDF